// app/schemas/registerSchema.ts
import { z } from "zod";

export const registerSchema = z.object({
  fullName: z.string().min(2, "Name must be at least 2 characters"),
  email: z.string().email("Please enter a valid email address"),
  password: z
    .string()
    .min(6, "Password must be at least 6 characters")
    .max(30, "Password must not exceed 30 characters"),
  phone: z.string().optional(),
  location: z.string().optional(),
  bio: z.string().optional(),
});

export type RegisterSchemaType = z.infer<typeof registerSchema>;